edit listing api route

+++ api/controllers/listing.controller.js
@@ -16,11 +16,41 @@ export const deleteListing = async (req, res, next) => {
   if (!listing) {
     return res.status(404).json({ message: "Listing not found" });
   }
+  if (req.user.id !== listing.userRef) {
+    return res
+      .status(403)
+      .json({ message: "You can only delete your listings" });
+  }
 
   try {
     await Listing.findByIdAndDelete(req.params.id);
 
     return res.status(200).json({ message: "Listing has been deleted" });
+  } catch (err) {
+    next(err);
+  }
+};
+
+export const updateListing = async (req, res, next) => {
+  const listing = await Listing.findById(req.params.id);
+
+  if (!listing) {
+    return res.status(404).json({ message: "Listing not found" });
+  }
+
+  if (req.user.id !== listing.userRef) {
+    return res
+      .status(403)
+      .json({ message: "You can only update your listings" });
+  }
+  try {
+    const updatedListing = await Listing.findByIdAndUpdate(
+      req.params.id,
+      req.body,
+      { new: true }
+    );
+
+    return res.status(200).json(updatedListing);
   } catch (err) {
     next(err);
   }
